+++ lib/puppet_x/voxpupuli/corosync/provider/crmsh.rb
@@ -1,5 +1,13 @@
-require 'puppet_x/voxpupuli/corosync/provider'
-require 'puppet_x/voxpupuli/corosync/provider/cib_helper'
+begin
+  require 'puppet_x/voxpupuli/corosync/provider'
+  require 'puppet_x/voxpupuli/corosync/provider/cib_helper'
+rescue LoadError
+  require 'pathname' # WORKAROUND #14073, #7788 and SERVER-973
+  corosync = Puppet::Module.find('corosync', Puppet[:environment].to_s)
+  raise(LoadError, "Unable to find corosync module in modulepath #{Puppet[:basemodulepath] || Puppet[:modulepath]}") unless corosync
+  require File.join corosync.path, 'lib/puppet_x/voxpupuli/corosync/provider'
+  require File.join corosync.path, 'lib/puppet_x/voxpupuli/corosync/provider/cib_helper'
+end
 
 class PuppetX::Voxpupuli::Corosync::Provider::Crmsh < PuppetX::Voxpupuli::Corosync::Provider::CibHelper
   # Yep, that's right we are parsing XML...FUN! (It really wasn't that bad)